Hide VT cursor before displaying SHR splash
Hi,

attached is a patch to hide the VT cursor before displaying SHR splash.

Don't know about you, but that cursor destroying the splash bitmap annoys
me. :)

There are of course other ways to do it (i.e. "tput civis") that might be
more closs-platform/terminal, but I couldn't make it work in that init
script. But that's probably not an issue, as we're always running on linux
term.

Regards,

zub

Signed-off-by: Sebastian Krzyszkowiak <seba.dos1@gmail.com>